style: Moving dtos implementations to a separate package
