+++ src/main/java/br/devrafaelsoares/storeapirestful/domain/product/dto/ProductResponse.java
@@ -1,6 +1,7 @@
-package br.devrafaelsoares.storeapirestful.domain.product;
+package br.devrafaelsoares.storeapirestful.domain.product.dto;
 
 import br.devrafaelsoares.storeapirestful.domain.category.Category;
+import br.devrafaelsoares.storeapirestful.domain.product.Product;
 
 import java.io.Serializable;
 import java.util.UUID;
