+++ src/main/java/br/devrafaelsoares/storeapirestful/domain/product/dto/ProductUpdate.java
@@ -1,4 +1,4 @@
-package br.devrafaelsoares.storeapirestful.domain.product;
+package br.devrafaelsoares.storeapirestful.domain.product.dto;
 
 public interface ProductUpdate {
